requirements.txt
streamlit
pandas
pymysql
python-dotenv
streamlit-aggrid

